% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ExistenceTest_Pixels}
\alias{ExistenceTest_Pixels}
\title{ExistenceTest_Pixels}
\usage{
ExistenceTest_Pixels(
  projectName,
  ERA,
  CMIP5,
  input_dir,
  agentNoLookup,
  PRM_Grid,
  PRM_RCPs,
  output
)
}
\arguments{
\item{projectName}{A string shows project name}

\item{ERA}{A character vector contains met file names. Returned from `listMets`}

\item{CMIP5}{A character vector contains met file names. Returned from `listMets`}

\item{agentNoLookup}{A sting shows the path to the agent number file with GPS location}

\item{PRM_Grid}{File path to the required agent number}

\item{PRM_RCPs}{File path to the climate models}

\item{output}{File path to output a missing list if any met files not there}
}
\value{
a data.frame shows the missing pixels
}
\description{
ExistenceTest_Pixels
}
\seealso{
\code{\link{listMets}}
}
\author{
Jian(Frank) Liu
}
