% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filterGridCells}
\alias{filterGridCells}
\title{filterGridCells}
\usage{
filterGridCells(path, PRM_Grid, output)
}
\arguments{
\item{path}{path to the file reports missing met files}

\item{PRM_Grid}{path to the input grid cell}
}
\value{

}
\description{
Pixels along the coast or lake may be lack of raw data in the `netcdf`.
It might be worth to filter out those ones if there are still missing mets after regeneration.
}
