% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{FindIncompleteJobs}
\alias{FindIncompleteJobs}
\title{FindIncompleteJobs}
\usage{
FindIncompleteJobs(outdir, outpath)
}
\arguments{
\item{outdir}{Directory to where the simulation files hold}

\item{outpath}{path to a file to output the list of trouble simulation files.}
}
\value{

}
\description{
Find the failed and incomplete jobs after all simulations.
}
