% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAcquisition.R
\name{get_connected}
\alias{get_connected}
\title{get_connected}
\usage{
get_connected(
  projectName = Sys.getenv("projectName"),
  credspath = Sys.getenv("secret_path")
)
}
\arguments{
\item{projectName}{Character string for the project name. prefer to source
from evnvironmental variables}

\item{credspath}{Character string to the credential file which has the
authentification to access PostgreSQL dbs for projects.}
}
\value{
A PostgreSQL db connection.
}
\description{
Build the connection between R session and PostgreSQL dbs
}
\examples{
get_connected()
}
