% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAcquisition.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(
  variables = NULL,
  dbconnection = connection,
  path = Sys.getenv("path")
)
}
\arguments{
\item{variables}{A named list. The element name must be the desired variable
names and the values of the element are for filtering the DB. Default is to
retrieve the entire \strong{Simulation} table. See details for constructing
 such named list.}

\item{dbconnection}{A PostgreSQL connection.}

\item{path}{A character string to show where to save the retrieved data.}
}
\value{

}
\description{
Retrieve selected/filtered data from the \strong{Simulation}
 table. The retrieved data will be saved in the project folder in a
 \strong{RDS} format.
}
\seealso{
\code{\link{get_table}}
}
