% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAcquisition.R
\name{get_info}
\alias{get_info}
\title{get_info}
\usage{
get_info(projectName = Sys.getenv("projectName"), dbconnection = connection)
}
\arguments{
\item{projectName}{Character string for the project name. prefer to source
from evnvironmental variables}

\item{dbconnection}{The connection to access the project db. The connection
will be closed once the function is called.}
}
\value{
A print out description about the db and a data.frame which contains
  the factor table of the db.
}
\description{
Basic description about the project db. Information such as how
  many tables, variables, number of pixels etc.
}
\examples{
get_info(projectName = Sys.getenv("projectName"), dbconnection = get_connected())
}
\seealso{
\code{\link{get_connected}}
}
