% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAcquisition.R
\name{get_table}
\alias{get_table}
\title{get_table}
\usage{
get_table(connection = get_connected(), table = "Variable")
}
\arguments{
\item{connection}{A PostgreSQL connection.}

\item{table}{Charater string to state which table is needed.}
}
\value{
A data.table that contains the data from the desired table in DB.
}
\description{
Retrieve the information from a table in the project DB.
}
\examples{
get_table(get_connected(), table = "Variable")
}
\seealso{
\code{\link{get_connected}}
}
