% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{listMets}
\alias{listMets}
\title{listMets}
\usage{
listMets(input_dir, historical = TRUE)
}
\arguments{
\item{input_dir}{A string to show the directory where the met files are}

\item{historical}{Logical. Default is TRUE, which ERA will be listed.}
}
\value{
vector has full met file names
}
\description{
call command line `find` to list all existing met files.
}
\examples{

}
