% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mvExistingApsimxFiles}
\alias{mvExistingApsimxFiles}
\title{mvExistingApsimxFiles}
\usage{
mvExistingApsimxFiles(outdir)
}
\value{

}
\description{
Move the existing `Temp_ApsimFiles` into another folder with date.
}
