% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ReSubmit}
\alias{ReSubmit}
\title{ReSubmit}
\usage{
ReSubmit(file, projectName, apsimversion)
}
\arguments{
\item{file}{The path to the file contains the failed or incomplete apsimx names.}
}
\value{

}
\description{
Re-submit the jobs from a txt file
}
\seealso{
[AtlasEDA::FindIncompleteJobs()]
}
