% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAcquisition.R
\name{get_graph}
\alias{get_graph}
\title{get_graph}
\usage{
get_graph(DT, plot_var, group_var)
}
\arguments{
\item{DT}{}
}
\value{

}
\description{
Experimenting. Might not necessary. Plot box_plot for all
 variables in the \strong{Variable} table by the value group.
}
